Replace vh as font size unit with vw in mobile
Why use viewport-width instead of viewport-height, as was previously
done? For one, on iOS Safari 7+, the viewport height changes as you
scroll up and down, due to the animation of the address bar. Thus,
using vh for font size means that as the user scrolls, the layout of
the page will change. This creates a rather unsettling effect, and
replacing vh with vw will fix this as the width of the viewport is
constant. Second, it creates a more stable layout. No matter the size
of the viewport, text flow will roughly stay the same, especially in
portrait orientation (aka de facto phone use mode). There are other
reasons, but they are minor compared to these two.